requirements.txt
hyperliquid-python-sdk>=0.4.0
eth-account>=0.11.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
